fixed to normal

+++ card.c
@@ -1,6 +1,6 @@
 #include "card.h"
 
-int8_t getValue(card *c)
+int getValue(card *c)
 {
     return c->value;
 }
